+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.6.6...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.6.7...HEAD) - TBD
+
+## [4.6.7](https://github.com/schemathesis/schemathesis/compare/v4.6.6...v4.6.7) - 2025-12-04
 
 ### :wrench: Changed